requirements.txt
Flask==2.3.0
requests==2.31.0
pydantic==2.0.0
crewai==0.1.0
crewai-tools==0.1.0
langchain-openai==0.1.0
python-dotenv==1.0.0
